+++ src/TM1638plus.h
@@ -19,6 +19,7 @@
 
 #include "TM1638plus_common.h"
 
+
 class TM1638plus  {
 
 public:
@@ -70,32 +71,39 @@ public:
 	// i.e 0b01000001 will set g and a on. 
 	void display7Seg(uint8_t position, uint8_t value);
 	
-	//Display an integer and leading zeros optional
-	void displayIntNum(unsigned long number, boolean leadingZeros = true);
+	// Display an integer and leading zeros optional
+	// Param 1 :: integer to display 2^32 
+	// Param 2 :: bool leading zeros , true on , false off
+	// Param 3 :: enum text alignment , left or right alignment
+	void displayIntNum(unsigned long number, boolean leadingZeros = true, AlignTextType_e = TMAlignTextLeft);
 	
-	//Divides the display into two nibbles and displays a Decimal number in each.
-	//takes in two numbers 0-9999 for each nibble ,  and byte for decimal point display,
-	//and leading zeros optional
-	void DisplayDecNumNibble(uint16_t numberUpper, uint16_t numberLower, boolean leadingZeros = true);
+	// Divides the display into two nibbles and displays a Decimal number in each.
+	// takes in two numbers 0-9999 for each nibble ,
+	// and leading zeros optional
+	// Param 1 :: upper nibble integer 2^16 
+	// Param 2 :: lower nibble integer 2^16 
+	// Param 3 :: bool leading zeros , true on , false off
+	// Param 4 :: enum text alignment , left or right alignment
+	void DisplayDecNumNibble(uint16_t numberUpper, uint16_t numberLower, boolean leadingZeros = true, AlignTextType_e = TMAlignTextLeft);
 	
-	  // Set the LEDs. passed one  16bit integer.
-	  // MODEL 3:
-	  //MSB byte for the green LEDs, LS byte for the red LEDs (0xgreenred) 
-	  //ie. 0xE007   1110 0000 0000 0111   results in L8-L0  GGGX XRRR, NOTE L8 is RHS on display
-	  // MODEL 1:
-	  // MSB byte 1 for  red LED , LSB byte n/a set to 0x00 (0xleds, 0xXX)
-	  //i.e 0xF100  1111 0000 L8-L0 RRRRXXX0 NOTE  L8 is RHS on display
-    void setLEDs(uint16_t greenred);
-    
-    // Set an LED, pass it LED position 0-7 and value 0 or 1 , L1-L8
+	// Set the LEDs. passed one  16bit integer.
+	// MODEL 3:
+	//MSB byte for the green LEDs, LS byte for the red LEDs (0xgreenred) 
+	//ie. 0xE007   1110 0000 0000 0111   results in L8-L0  GGGX XRRR, NOTE L8 is RHS on display
+	// MODEL 1:
+	// MSB byte 1 for  red LED , LSB byte n/a set to 0x00 (0xleds, 0xXX)
+	//i.e 0xF100  1111 0000 L8-L0 RRRRXXX0 NOTE  L8 is RHS on display
+	void setLEDs(uint16_t greenred);
+ 
+	// Set an LED, pass it LED position 0-7 and value 0 or 1 , L1-L8
 	void setLED(uint8_t position, uint8_t value);
 	
 private:
 		uint8_t _STROBE_IO;
 		uint8_t _DATA_IO;
 		uint8_t _CLOCK_IO;
-		 void sendCommand(uint8_t value);
-        void sendData(uint8_t  data);
+		void sendCommand(uint8_t value);
+		void sendData(uint8_t  data);
 		//This is used when running high freq CPU 
 		bool _HIGH_FREQ = false;
 		TM1638plus_common  TM_common;