+++ extra/CHANGELOG.md
@@ -52,3 +52,11 @@
 * Version 1.8.0 November 2021 
 	* Minor update 
 	* Some users (two) have reported the high frequency shiftIn  function does not work fully with some micro-controller boards at high frequency(ESP32 240Mhz) and the fix is minor adjustment to sequence in said function. Never saw the issue in my testing. The change makes  the function sequence similar to the official Arduino shiftIn function structure in the Arduino Core (wiring.shift.c) ,  See issue 16 on github for full details.
+
+* Version 1.9.0 November 2022
+	* Change made to Model 1/3 displayHex method as per pull request #19.
+	* Right aligned text mode added to (as per pull request #17)
+		* Model 1/3,  methods displayIntNum and DisplayDecNumNibble
+		* Model  2 , methods DisplayDecNum, DisplayDecNumNibble & DisplayHexNum
+		* Previously it was just Left aligned text and added leading zeros.
+